backend/requirements.txt
flask
dotenv
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
bcrypt
